chore(hog-charts): break BarChart into focused modules
BarChart.tsx had grown to 643 lines, mixing component orchestration with
canvas-drawing logic, config resolution, and click routing. Following the
pattern the LineChart family already uses (pure helpers extracted into
`utils/` siblings with co-located tests) and the hog-charts CONTRIBUTING.md
layering guidance, split the chart-specific logic out:

- `core/bar-layout.ts` — band-center geometry helpers (`bandCenter`,
  `groupedBarCenter`), the documented home for bar geometry.
- `utils/bar-config.ts` — shadow + wrapper-min-height resolution, with tests.
- `utils/draw-bar-chart.ts` — the static and hover canvas passes as pure
  functions; the component keeps only the hover-fade bookkeeping.
- `utils/resolve-clicked-bar-series.ts` — the pure click-routing function.

BarChart.tsx drops from 643 to 342 lines and now reads as a thin
orchestration layer, matching LineChart's altitude. No behavior change.

Generated-By: PostHog Code
Task-Id: 6cb78494-db76-47c8-95d0-598be8b523f0
