+++ frontend/src/lib/hog-charts/charts/BarChart/BarChart.tsx
@@ -1,25 +1,12 @@
-import * as d3 from 'd3'
 import React, { useCallback, useMemo, useRef } from 'react'
 
-import { type BarChartPrivate, computeBarTrackRect, computeSeriesBars } from '../../core/bar-layout'
-import {
-    BAR_TRACK_HOVER_ALPHA,
-    type BarRect,
-    type BarShadow,
-    drawBarHighlight,
-    drawBars,
-    drawBarTracks,
-    drawGrid,
-    type DrawContext,
-} from '../../core/canvas-renderer'
+import { type BarChartPrivate, bandCenter, groupedBarCenter } from '../../core/bar-layout'
 import { Chart } from '../../core/Chart'
 import { ChartErrorBoundary } from '../../core/ChartErrorBoundary'
-import { DEFAULT_MARGINS, X_AXIS_TITLE_MARGIN } from '../../core/hooks/useChartMargins'
 import { useLatest } from '../../core/hooks/useLatest'
 import {
     buildSegmentResolveValue,
     buildStackedPositionValue,
-    type BarScaleSet,
     computeDivergingStackData,
     computePercentStackData,
     computeStackData,
@@ -29,7 +16,6 @@ import {
 } from '../../core/scales'
 import type {
     BarChartConfig,
-    BarsConfig,
     ChartDimensions,
     ChartDrawArgs,
     ChartScales,
@@ -42,33 +28,15 @@ import type {
     TooltipContext,
 } from '../../core/types'
 import { DEFAULT_Y_AXIS_ID } from '../../core/types'
-import { computeVisibleXLabels } from '../../overlays/AxisLabels'
 import { BarTooltip } from './BarTooltip'
+import { computeWrapperMinHeight, HORIZONTAL_MIN_BAND_SIZE_DEFAULT } from './utils/bar-config'
 import {
-    type BarLayout,
-    barContainsPointOnBandAxis,
-    cursorOutsideBarFillExtent,
-    findVisibleStackedSegment,
-    iterBarsAtCursor,
-    isStackedLayout,
-} from './utils/bars-under-cursor'
-
-function bandCenter(scales: BarChartPrivate['__barChart'], label: string): number | undefined {
-    const start = scales.band(label)
-    return start == null ? undefined : start + scales.band.bandwidth() / 2
-}
-
-/** Center of a specific series's bar within a band. Used by overlays (e.g. annotations)
- *  to anchor on the current-period bar in compare-against-previous grouped layouts.
- *  Returns undefined when the layout isn't grouped or the series isn't in the group scale. */
-function groupedBarCenter(scales: BarChartPrivate['__barChart'], label: string, seriesKey: string): number | undefined {
-    const start = scales.band(label)
-    const groupOffset = scales.group?.(seriesKey)
-    if (start == null || groupOffset == null) {
-        return undefined
-    }
-    return start + groupOffset + (scales.group?.bandwidth() ?? 0) / 2
-}
+    type BarChartDrawConfig,
+    drawBarChartStatic,
+    drawBarHoverItems,
+    resolveBarHoverItems,
+} from './utils/draw-bar-chart'
+import { resolveClickedBarSeries } from './utils/resolve-clicked-bar-series'
 
 export interface BarChartProps<Meta = unknown> {
     series: Series<Meta>[]
@@ -82,24 +50,6 @@ export interface BarChartProps<Meta = unknown> {
     dataAttr?: string
     children?: React.ReactNode
     onError?: (error: Error, info: React.ErrorInfo) => void
-}
-
-// Negative offsetY casts the shadow upward onto the visible track above the bar.
-const DEFAULT_BAR_SHADOW: BarShadow = { color: 'rgba(0,0,0,0.30)', blur: 12, offsetY: -4 }
-
-// Horizontal floor: each row gets at least this much px so tick labels don't crush; wrapper scrolls.
-const HORIZONTAL_MIN_BAND_SIZE_DEFAULT = 24
-// Reserve room for chart-edge margins + worst-case x-axis title margin (matches useChartMargins).
-const HORIZONTAL_CHART_MARGIN_PX = DEFAULT_MARGINS.top + DEFAULT_MARGINS.bottom + X_AXIS_TITLE_MARGIN
-
-function resolveBarShadow(barShadow: BarsConfig['shadow']): BarShadow | undefined {
-    if (barShadow === true) {
-        return DEFAULT_BAR_SHADOW
-    }
-    if (barShadow === false || barShadow == null) {
-        return undefined
-    }
-    return barShadow
 }
 
 export function BarChart<Meta = unknown>({ onError, ...rest }: BarChartProps<Meta>): React.ReactElement {
@@ -140,16 +90,10 @@ function BarChartInner<Meta = unknown>({
     const isHorizontal = axisOrientation === 'horizontal'
 
     const resolvedMinBandSize = minBandSize ?? (isHorizontal ? HORIZONTAL_MIN_BAND_SIZE_DEFAULT : 0)
-    const wrapperMinHeight = useMemo(() => {
-        if (!isHorizontal || resolvedMinBandSize <= 0) {
-            return undefined
-        }
-        const uniqueBands = new Set(labels).size
-        if (uniqueBands === 0) {
-            return undefined
-        }
-        return uniqueBands * resolvedMinBandSize + HORIZONTAL_CHART_MARGIN_PX
-    }, [isHorizontal, resolvedMinBandSize, labels])
+    const wrapperMinHeight = useMemo(
+        () => computeWrapperMinHeight({ isHorizontal, resolvedMinBandSize, labels }),
+        [isHorizontal, resolvedMinBandSize, labels]
+    )
 
     const stackedData = useMemo((): Map<string, StackedBand> | undefined => {
         if (barLayout === 'percent') {
@@ -273,219 +217,57 @@ function BarChartInner<Meta = unknown>({
         [yScaleType, barLayout, axisOrientation, stackedData, isHorizontal, divergingStack, maxBandRange, bandPadding]
     )
 
-    const drawStatic = useCallback(
-        ({ ctx, dimensions, scales, series: coloredSeries, labels: drawLabels, theme }: ChartDrawArgs) => {
-            const d3Scales = (scales._private as BarChartPrivate | undefined)?.__barChart
-            if (!d3Scales) {
-                return
-            }
-
-            const baseDrawCtx: DrawContext = {
-                ctx,
-                dimensions,
-                xScale: (label: string) => bandCenter(d3Scales, label),
-                yScale: d3Scales.value,
-                labels: drawLabels,
-            }
-
-            if (showGrid) {
-                // Align cross-axis grid with visible category labels, not every band.
-                let categoryTicks: number[] = []
-                if (isHorizontal) {
-                    for (const label of drawLabels) {
-                        const coord = bandCenter(d3Scales, label)
-                        if (coord != null && isFinite(coord)) {
-                            categoryTicks.push(coord)
-                        }
-                    }
-                } else {
-                    categoryTicks = computeVisibleXLabels(
-                        drawLabels,
-                        (label) => bandCenter(d3Scales, label),
-                        xTickFormatter
-                    ).map((entry) => entry.x)
-                }
-                drawGrid(baseDrawCtx, {
-                    gridColor: theme.gridColor,
-                    orientation: isHorizontal ? 'horizontal' : 'vertical',
-                    categoryTicks,
-                })
-            }
-
-            const seriesBars = coloredSeries
-                .filter((s) => !s.visibility?.excluded)
-                .map((s) => {
-                    const axisId = s.yAxisId ?? DEFAULT_Y_AXIS_ID
-                    const bars = computeSeriesBars({
-                        series: s,
-                        labels: drawLabels,
-                        scales: d3Scales,
-                        layout: barLayout,
-                        isHorizontal,
-                        stackedBand: stackedData?.get(s.key),
-                        isTopOfStack: topStackedKeyByAxis.get(axisId) === s.key,
-                    }).filter((b): b is BarRect => b !== null)
-                    return { series: s, bars }
-                })
-
-            // Tracks are a separate pass so a later series' full-height track can't paint
-            // over an earlier series' bar. Track is "share of a whole" semantics — only
-            // meaningful for grouped layouts; in stacked/percent every layer would paint
-            // a full-height track over the same band with the wrong corners.
-            if (barTrack && barLayout === 'grouped') {
-                const [axisStart = 0, axisEnd = 0] = d3Scales.value.range()
-                for (const { series: s, bars } of seriesBars) {
-                    const tracks = bars.map((b) => computeBarTrackRect(b, axisStart, axisEnd, isHorizontal))
-                    drawBarTracks(baseDrawCtx, s, tracks, barCornerRadius)
-                }
-            }
-
-            const resolvedShadow = resolveBarShadow(barShadow)
-            // Clip to plot area so a 100% bar's upward shadow doesn't bleed past the chart edge.
-            if (resolvedShadow) {
-                ctx.save()
-                ctx.beginPath()
-                ctx.rect(dimensions.plotLeft, dimensions.plotTop, dimensions.plotWidth, dimensions.plotHeight)
-                ctx.clip()
-                ctx.shadowColor = resolvedShadow.color
-                ctx.shadowBlur = resolvedShadow.blur
-                ctx.shadowOffsetX = resolvedShadow.offsetX ?? 0
-                ctx.shadowOffsetY = resolvedShadow.offsetY ?? 0
-            }
-            for (const { series: s, bars } of seriesBars) {
-                drawBars(baseDrawCtx, s, bars, barCornerRadius)
-            }
-            if (resolvedShadow) {
-                ctx.restore()
-            }
-        },
-        [
-            showGrid,
-            stackedData,
+    const drawConfig = useMemo<BarChartDrawConfig>(
+        () => ({
             barLayout,
             isHorizontal,
+            stackedData,
             topStackedKeyByAxis,
             barCornerRadius,
             barTrack,
+            showGrid,
+            xTickFormatter,
+            barShadow,
+        }),
+        [
+            barLayout,
+            isHorizontal,
+            stackedData,
+            topStackedKeyByAxis,
+            barCornerRadius,
+            barTrack,
+            showGrid,
             xTickFormatter,
             barShadow,
         ]
     )
 
+    const drawStatic = useCallback((args: ChartDrawArgs) => drawBarChartStatic(args, drawConfig), [drawConfig])
+
     // Restart the fade on bar → track moves (same hoverIndex, different visible state).
     const lastHoverKeyRef = useRef<string | null>(null)
 
     const drawHover = useCallback(
-        ({
-            ctx,
-            scales,
-            series: coloredSeries,
-            labels: drawLabels,
-            hoverIndex,
-            hoverPosition,
-            hoverProgress,
-            resetHoverFade,
-        }: ChartDrawArgs): DrawHoverResult => {
-            const d3Scales = (scales._private as BarChartPrivate | undefined)?.__barChart
-            if (!d3Scales || hoverIndex < 0) {
+        (args: ChartDrawArgs): DrawHoverResult => {
+            const resolved = resolveBarHoverItems(args, drawConfig)
+            if (!resolved) {
                 lastHoverKeyRef.current = null
                 return false
             }
-            const hoveredLabel = drawLabels[hoverIndex]
-            const [trackAxisStart = 0, trackAxisEnd = 0] = barTrack ? d3Scales.value.range() : []
-            // Key includes bar-vs-track per series so bar → track moves at the same
-            // hoverIndex still trigger a fade restart.
-            type DrawItem = { series: ResolvedSeries; bar: BarRect; isTrackHighlight: boolean }
-            const items: DrawItem[] = []
-            let composition = ''
-            // Stacked: clip the highlight to the visible slice so hover only changes shade,
-            // never z-order. Grouped keeps band-axis containment for cursor-above-bar.
-            const stackedHighlight = isStackedLayout(barLayout)
-            if (stackedHighlight && hoverPosition) {
-                const visible = findVisibleStackedSegment({
-                    series: coloredSeries,
-                    labels: drawLabels,
-                    hoveredLabel,
-                    cursor: hoverPosition,
-                    scales: d3Scales,
-                    layout: barLayout,
-                    isHorizontal,
-                    stackedData,
-                    topStackedKeyByAxis,
-                })
-                if (visible) {
-                    const visibleExtent = isHorizontal ? visible.bar.width : visible.bar.height
-                    const { nextSmallerExtent } = visible
-                    const baselinePx = isHorizontal ? visible.bar.x : visible.bar.y + visible.bar.height
-                    const clippedExtent = Math.max(0, visibleExtent - nextSmallerExtent)
-                    const clipped: BarRect = isHorizontal
-                        ? { ...visible.bar, x: baselinePx + nextSmallerExtent, width: clippedExtent }
-                        : { ...visible.bar, y: baselinePx - visibleExtent, height: clippedExtent }
-                    items.push({ series: visible.series, bar: clipped, isTrackHighlight: false })
-                    composition += 'b'
-                }
-            } else {
-                for (const { series: s, bar } of iterBarsAtCursor<ResolvedSeries>({
-                    series: coloredSeries,
-                    label: hoveredLabel,
-                    dataIndex: hoverIndex,
-                    scales: d3Scales,
-                    layout: barLayout,
-                    isHorizontal,
-                    stackedData,
-                    topStackedKeyByAxis,
-                })) {
-                    if (hoverPosition && !barContainsPointOnBandAxis(bar, hoverPosition, isHorizontal)) {
-                        continue
-                    }
-                    const isTrackHighlight =
-                        barTrack === true &&
-                        barLayout === 'grouped' &&
-                        hoverPosition != null &&
-                        cursorOutsideBarFillExtent(bar, hoverPosition, isHorizontal)
-                    items.push({ series: s, bar, isTrackHighlight })
-                    composition += isTrackHighlight ? 't' : 'b'
-                }
-            }
-            if (items.length === 0) {
-                lastHoverKeyRef.current = null
-                return false
-            }
-            const currentKey = `${hoverIndex}:${composition}`
-            let alpha = hoverProgress
+            const { items, composition, trackAxisRange } = resolved
+            const currentKey = `${args.hoverIndex}:${composition}`
+            let alpha = args.hoverProgress
             if (currentKey !== lastHoverKeyRef.current) {
-                alpha = resetHoverFade()
+                alpha = args.resetHoverFade()
                 lastHoverKeyRef.current = currentKey
             }
-            ctx.save()
-            ctx.globalAlpha = alpha
-            for (const { series: s, bar, isTrackHighlight } of items) {
-                if (isTrackHighlight) {
-                    const parsed = d3.color(s.color)
-                    // Always translucent — `s.color` direct would paint an opaque full-height
-                    // block if d3 can't parse the color.
-                    let trackColor: string
-                    if (parsed) {
-                        parsed.opacity = BAR_TRACK_HOVER_ALPHA
-                        trackColor = parsed.toString()
-                    } else {
-                        trackColor = `rgba(0,0,0,${BAR_TRACK_HOVER_ALPHA})`
-                    }
-                    drawBarHighlight(
-                        ctx,
-                        computeBarTrackRect(bar, trackAxisStart, trackAxisEnd, isHorizontal),
-                        trackColor,
-                        barCornerRadius
-                    )
-                } else {
-                    const highlightColor = d3.color(s.color)?.darker(0.6).toString() ?? s.color
-                    drawBarHighlight(ctx, bar, highlightColor, barCornerRadius)
-                }
-            }
-            ctx.restore()
+            args.ctx.save()
+            args.ctx.globalAlpha = alpha
+            drawBarHoverItems(args.ctx, items, { barCornerRadius, isHorizontal, trackAxisRange })
+            args.ctx.restore()
             return true
         },
-        [stackedData, barLayout, isHorizontal, topStackedKeyByAxis, barCornerRadius, barTrack]
+        [drawConfig, barCornerRadius, isHorizontal]
     )
 
     // Show each series's own segment value (resolveValue) but anchor the tooltip/value labels
@@ -557,87 +339,4 @@ function BarChartInner<Meta = unknown>({
             {chart}
         </div>
     )
-}
-
-/** Rewrites the click payload to the bar series actually under the cursor. The base payload
- *  always points at the first series in the band; this picks the right one per layout:
- *   - grouped: the series whose sub-band column the cursor is over — band axis only, so a
- *     click above a short bar (or on its track) still resolves to that column.
- *   - stacked/percent: the segment whose rect contains the cursor on the value axis, walking
- *     every dataIndex in the band so sparse-overlap segments route correctly, and re-reading
- *     the value at that segment's own dataIndex.
- *  Pure so the routing is unit-testable; returns `null` to pass `clickData` through unchanged. */
-export function resolveClickedBarSeries<Meta>({
-    clickData,
-    d3Scales,
-    barLayout,
-    isHorizontal,
-    stackedData,
-    topStackedKeyByAxis,
-    series,
-    labels,
-}: {
-    clickData: PointClickData<Meta>
-    d3Scales: BarScaleSet
-    barLayout: BarLayout
-    isHorizontal: boolean
-    stackedData: Map<string, StackedBand> | undefined
-    topStackedKeyByAxis: Map<string, string>
-    series: Series<Meta>[]
-    labels: readonly string[]
-}): PointClickData<Meta> | null {
-    const { cursor, label, dataIndex, crossSeriesData } = clickData
-    if (!cursor) {
-        return null
-    }
-    const rewrite = (hitSeries: Series<Meta>, value: number, hitDataIndex: number): PointClickData<Meta> => ({
-        ...clickData,
-        dataIndex: hitDataIndex,
-        series: hitSeries,
-        value,
-        seriesIndex: series.findIndex((s) => s.key === hitSeries.key),
-    })
-
-    if (barLayout === 'grouped') {
-        for (const { series: s, bar } of iterBarsAtCursor({
-            series: crossSeriesData.map((d) => d.series),
-            label,
-            dataIndex,
-            scales: d3Scales,
-            layout: barLayout,
-            isHorizontal,
-            topStackedKeyByAxis,
-        })) {
-            if (!barContainsPointOnBandAxis(bar, cursor, isHorizontal)) {
-                continue
-            }
-            const hit = crossSeriesData.find((d) => d.series.key === s.key)
-            return hit ? rewrite(hit.series, hit.value, dataIndex) : null
-        }
-        return null
-    }
-
-    const visible = findVisibleStackedSegment({
-        series: crossSeriesData.map((d) => d.series),
-        labels,
-        hoveredLabel: label,
-        cursor,
-        scales: d3Scales,
-        layout: barLayout,
-        isHorizontal,
-        stackedData,
-        topStackedKeyByAxis,
-    })
-    if (!visible) {
-        return null
-    }
-    const hit = crossSeriesData.find((d) => d.series.key === visible.series.key)
-    if (!hit) {
-        return null
-    }
-    // Re-read value at the visible segment's own dataIndex — `hit.value` was resolved at the
-    // band's dataIndex, which is a sparse-zero cell for the visible series.
-    const raw = hit.series.data[visible.dataIndex]
-    const resolvedValue = typeof raw === 'number' && Number.isFinite(raw) ? raw : hit.value
-    return rewrite(hit.series, resolvedValue, visible.dataIndex)
 }
